Change onCleanup regex
The current regex only fixes the last ampersand within modx tags with multiple properties being set. This takes a different approach, not attempting to focus in on content within opening/closing double brackets but rather on the parameter format itself (which is highly unlikely to be used in normal text). This would also allow for ampersands within nested tags to be easily converted too, sidestepping the need for an overly-complex regex.